Upgrading to CRA level 3.3.1

+++ packages/borderline-devutils/bin/borderline-devutils.js
@@ -1,10 +1,7 @@
-#!/usr/bin/env node
-/* eslint no-console: "off" */
 // Makes the script crash on unhandled rejections instead of silently
 // ignoring them. In the future, promise rejections that are not handled will
 // terminate the Node.js process with a non-zero exit code.
 process.on('unhandledRejection', err => {
-    console.error(err.stack);
     throw err;
 });
 
@@ -17,38 +14,31 @@ const scriptIndex = args.findIndex(
 const script = scriptIndex === -1 ? args[0] : args[scriptIndex];
 const nodeArgs = scriptIndex > 0 ? args.slice(0, scriptIndex) : [];
 
-switch (script) {
-    case 'build':
-    case 'publish':
-    case 'start':
-    case 'test': {
-        const result = spawn.sync(
-            'node',
-            nodeArgs
-                .concat(require.resolve('../scripts/' + script))
-                .concat(args.slice(scriptIndex + 1)),
-            { stdio: 'inherit' }
-        );
-        if (result.signal) {
-            if (result.signal === 'SIGKILL') {
-                console.log(
-                    'The build failed because the process exited too early. ' +
-                    'This probably means the system ran out of memory or someone called ' +
-                    '`kill -9` on the process.'
-                );
-            } else if (result.signal === 'SIGTERM') {
-                console.log(
-                    'The build failed because the process exited too early. ' +
-                    'Someone might have called `kill` or `killall`, or the system could ' +
-                    'be shutting down.'
-                );
-            }
-            process.exit(1);
+if (['build', 'start', 'test'].includes(script)) {
+    const result = spawn.sync(
+        'node',
+        nodeArgs
+            .concat(require.resolve('../scripts/' + script))
+            .concat(args.slice(scriptIndex + 1)),
+        { stdio: 'inherit' }
+    );
+    if (result.signal) {
+        if (result.signal === 'SIGKILL') {
+            console.log(
+                'The build failed because the process exited too early. ' +
+                'This probably means the system ran out of memory or someone called ' +
+                '`kill -9` on the process.'
+            );
+        } else if (result.signal === 'SIGTERM') {
+            console.log(
+                'The build failed because the process exited too early. ' +
+                'Someone might have called `kill` or `killall`, or the system could ' +
+                'be shutting down.'
+            );
         }
-        process.exit(result.status);
-        break;
+        process.exit(1);
     }
-    default:
-        console.log(`Unknown command "${script}".`);
-        break;
+    process.exit(result.status);
+} else {
+    console.log('Unknown script "' + script + '".');
 }
